requirements.txt
# Core dependencies for LogBERT anomaly detection system
torch>=2.0.0
transformers>=4.42.0
datasets>=2.19.0
accelerate>=0.30.0

# Data processing and ML utilities
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Log processing and template mining
drain3>=0.9.11
regex>=2023.6.3

# Web service and API
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Workflow orchestration
prefect==2.14.0

# Data storage and serialization
pyarrow>=12.0.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.0.0

# Logging and monitoring
structlog>=23.1.0
colorlog>=6.7.0

# Visualization and analysis (optional)
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# OpenSearch integration (optional)
opensearch-py>=2.3.0

# HTTP client for external APIs
requests>=2.31.0
aiohttp>=3.8.0

# Date/time utilities
python-dateutil>=2.8.2

# Progress bars and utilities
tqdm>=4.65.0

# JSON processing
orjson>=3.9.0

# Security and hashing
cryptography>=41.0.0

# Performance monitoring
psutil>=5.9.0
memory-profiler>=0.60.0